chore(Deployment): :rocket: add GH pages configs
The poor performances of Netlify are more than we can handle, on first load the website take around 20 seconds to display anything, or we get instantly an error, but sometimes even after first load, some ressources are not loading. we were also promised better performances using their CDN, but no. it's slower. but at least the images works in the blog. that will have to go too.